+++ src/tepkit/functions/vasp/poscar_tools.py
@@ -1,0 +1,6 @@
+from tepkit.io.vasp import Poscar
+
+
+def get_poscar_volume_cli(poscar="POSCAR"):
+    poscar = Poscar.from_file(poscar)
+    print(poscar.get_volume())